[Refactor] PhotoDetail Repo 분리
