+++ app/src/main/java/com/and04/naturealbum/data/repository/firebase/AlbumRepository.kt
@@ -7,9 +7,9 @@ import com.and04.naturealbum.data.dto.FirebaseLabel
 import com.and04.naturealbum.data.dto.FirebaseLabelResponse
 import com.and04.naturealbum.data.dto.FirebasePhotoInfo
 import com.and04.naturealbum.data.dto.FirebasePhotoInfoResponse
-import com.and04.naturealbum.data.repository.local.LocalDataRepository
 import com.and04.naturealbum.data.localdata.room.Label
 import com.and04.naturealbum.data.repository.local.LocalAlbumRepository
+import com.and04.naturealbum.data.repository.local.PhotoDetailRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
@@ -45,7 +45,7 @@ interface AlbumRepository {
 
 class AlbumRepositoryImpl @Inject constructor(
     private val firebaseDataSource: FirebaseDataSource,
-    private val localRepository: LocalDataRepository,
+    private val photoDetailRepository: PhotoDetailRepository,
     private val localAlbumRepository: LocalAlbumRepository,
 ) : AlbumRepository {
     override suspend fun getLabelsToList(uid: String): Result<List<FirebaseLabelResponse>> {
@@ -138,7 +138,7 @@ class AlbumRepositoryImpl @Inject constructor(
                                     firebaseDataSource.deleteUserLabel(uid, label)
                                 } else {
                                     val albumPresentFileName =
-                                        localRepository.getPhotoDetailById(albums[0].photoDetailId).fileName
+                                        photoDetailRepository.getPhotoDetailById(albums[0].photoDetailId).fileName
                                     val document =
                                         firebaseDataSource.getPhotoInfo(uid, albumPresentFileName)
                                     document.toObject(FirebasePhotoInfoResponse::class.java)
