+++ app/src/main/java/com/and04/naturealbum/di/FirebaseModule.kt
@@ -6,7 +6,7 @@ import com.and04.naturealbum.data.repository.firebase.AlbumRepositoryImpl
 import com.and04.naturealbum.data.repository.firebase.FriendRepository
 import com.and04.naturealbum.data.repository.firebase.UserRepository
 import com.and04.naturealbum.data.repository.local.LocalAlbumRepository
-import com.and04.naturealbum.data.repository.local.LocalDataRepository
+import com.and04.naturealbum.data.repository.local.PhotoDetailRepository
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
@@ -40,7 +40,7 @@ object FirebaseModule {
     @Singleton
     fun providerFireBaseRepository(
         firebaseDataSource: FirebaseDataSource,
-        localDataRepository: LocalDataRepository,
+        localDataRepository: PhotoDetailRepository,
         localAlbumRepository: LocalAlbumRepository
     ): AlbumRepository =
         AlbumRepositoryImpl(firebaseDataSource, localDataRepository, localAlbumRepository)
