+++ app/src/main/java/com/and04/naturealbum/di/RoomModule.kt
@@ -1,17 +1,17 @@
 package com.and04.naturealbum.di
 
 import android.content.Context
-import com.and04.naturealbum.data.repository.local.LocalDataRepository
-import com.and04.naturealbum.data.repository.local.LocalDataRepositoryImpl
 import com.and04.naturealbum.data.localdata.room.AlbumDao
 import com.and04.naturealbum.data.localdata.room.AppDatabase
 import com.and04.naturealbum.data.localdata.room.LabelDao
 import com.and04.naturealbum.data.localdata.room.PhotoDetailDao
 import com.and04.naturealbum.data.repository.local.LabelRepository
 import com.and04.naturealbum.data.repository.local.LocalAlbumRepository
-import com.and04.naturealbum.data.repository.local.impl.LocalAlbumRepositoryImpl
+import com.and04.naturealbum.data.repository.local.PhotoDetailRepository
 import com.and04.naturealbum.data.repository.local.SyncRepository
 import com.and04.naturealbum.data.repository.local.impl.LabelRepositoryImpl
+import com.and04.naturealbum.data.repository.local.impl.LocalAlbumRepositoryImpl
+import com.and04.naturealbum.data.repository.local.impl.PhotoDetailRepositoryImpl
 import com.and04.naturealbum.data.repository.local.impl.SyncRepositoryImpl
 import dagger.Module
 import dagger.Provides
@@ -53,7 +53,7 @@ object RoomModule {
     fun providerRepository(
         albumDao: AlbumDao,
         photoDetailDao: PhotoDetailDao
-    ): LocalDataRepository = LocalDataRepositoryImpl(albumDao, photoDetailDao)
+    ): PhotoDetailRepository = PhotoDetailRepositoryImpl(albumDao, photoDetailDao)
 
     @Singleton
     @Provides
