--- app/src/main/java/com/and04/naturealbum/ui/home/HomeViewModel.kt
@@ -1,7 +1,0 @@
-package com.and04.naturealbum.ui.home
-
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
-
-}
