+++ app/src/main/java/com/and04/naturealbum/ui/maps/MapScreenViewModel.kt
@@ -7,7 +7,7 @@ import com.and04.naturealbum.data.dto.FirebaseFriend
 import com.and04.naturealbum.data.repository.firebase.AlbumRepository
 import com.and04.naturealbum.data.repository.firebase.FriendRepository
 import com.and04.naturealbum.data.repository.local.LabelRepository
-import com.and04.naturealbum.data.repository.local.LocalDataRepository
+import com.and04.naturealbum.data.repository.local.PhotoDetailRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -17,7 +17,7 @@ import javax.inject.Inject
 
 @HiltViewModel
 class MapScreenViewModel @Inject constructor(
-    private val localRepository: LocalDataRepository,
+    private val photoDetailRepository: PhotoDetailRepository,
     private val albumRepository: AlbumRepository,
     private val friendRepository: FriendRepository,
     private val labelRepository: LabelRepository,
@@ -31,7 +31,7 @@ class MapScreenViewModel @Inject constructor(
 
     init {
         viewModelScope.launch {
-            val fetchPhotos = async { localRepository.getAllPhotoDetail() }
+            val fetchPhotos = async { photoDetailRepository.getAllPhotoDetail() }
             val fetchLabels = labelRepository.getLabels()
             myPhotos = fetchPhotos.await().toPhotoItems(fetchLabels)
             _photosByUid.emit(mapOf("" to myPhotos))
